remove dropzone pacakge
