--- packages/dropzone/package.json
@@ -1,151 +1,0 @@
-{
-  "name": "@uploadthing/dropzone",
-  "version": "0.4.1",
-  "type": "module",
-  "sideEffects": false,
-  "license": "MIT",
-  "exports": {
-    "./package.json": "./package.json",
-    "./core": {
-      "import": {
-        "types": "./core/index.d.ts",
-        "default": "./core/index.js"
-      },
-      "require": {
-        "types": "./core/index.d.cts",
-        "default": "./core/index.cjs"
-      }
-    },
-    "./react": {
-      "import": {
-        "types": "./react/index.d.ts",
-        "default": "./react/index.js"
-      },
-      "require": {
-        "types": "./react/index.d.cts",
-        "default": "./react/index.cjs"
-      }
-    },
-    "./solid": {
-      "import": {
-        "types": "./solid/index.d.ts",
-        "default": "./solid/index.js"
-      },
-      "require": {
-        "types": "./solid/index.d.cts",
-        "default": "./solid/index.cjs"
-      }
-    },
-    "./svelte": {
-      "import": {
-        "types": "./svelte/index.d.ts",
-        "default": "./svelte/index.js"
-      },
-      "require": {
-        "types": "./svelte/index.d.cts",
-        "default": "./svelte/index.cjs"
-      }
-    },
-    "./vue": {
-      "import": {
-        "types": "./vue/index.d.ts",
-        "default": "./vue/index.js"
-      },
-      "require": {
-        "types": "./vue/index.d.cts",
-        "default": "./vue/index.cjs"
-      }
-    }
-  },
-  "files": [
-    "core",
-    "react",
-    "solid",
-    "svelte",
-    "vue"
-  ],
-  "publishConfig": {
-    "access": "public"
-  },
-  "scripts": {
-    "lint": "eslint src --max-warnings 0",
-    "build": "bunchee --tsconfig tsconfig.build.json",
-    "clean": "git clean -xdf core react solid node_modules",
-    "dev": "bunchee -w --tsconfig tsconfig.build.json --no-clean",
-    "prepack": "bun ../../.github/replace-workspace-protocol.ts",
-    "typecheck": "tsc --noEmit"
-  },
-  "dependencies": {
-    "file-selector": "^0.6.0"
-  },
-  "devDependencies": {
-    "@types/react": "18.3.3",
-    "@uploadthing/eslint-config": "workspace:*",
-    "@uploadthing/tsconfig": "workspace:*",
-    "bunchee": "^5.2.1",
-    "eslint": "^8.57.0",
-    "react": "18.3.1",
-    "solid-js": "^1.8.15",
-    "svelte": "^4.2.12",
-    "typescript": "^5.5.2",
-    "vue": "^3.4.21"
-  },
-  "peerDependencies": {
-    "react": "^16.8.0 || ^17.0.0 || ^18.0.0",
-    "solid-js": "^1.7.11",
-    "svelte": "^4.2.12",
-    "vue": "^3.4.0"
-  },
-  "peerDependenciesMeta": {
-    "react": {
-      "optional": true
-    },
-    "solid-js": {
-      "optional": true
-    },
-    "svelte": {
-      "optional": true
-    },
-    "vue": {
-      "optional": true
-    }
-  },
-  "eslintConfig": {
-    "root": true,
-    "extends": [
-      "@uploadthing/eslint-config/base"
-    ],
-    "rules": {
-      "no-restricted-imports": [
-        "error",
-        {
-          "patterns": [
-            {
-              "group": [
-                "@uploadthing/dropzone",
-                "@uploadthing/dropzone/*"
-              ],
-              "message": "Use relative src imports instead"
-            }
-          ],
-          "paths": [
-            {
-              "name": "effect",
-              "message": "Use alias imports instead (import * as X from \"effect/X\")"
-            }
-          ]
-        }
-      ]
-    },
-    "overrides": [
-      {
-        "files": [
-          "src/react.tsx"
-        ],
-        "extends": [
-          "@uploadthing/eslint-config/react"
-        ]
-      }
-    ]
-  }
-}
