--- packages/dropzone/tsconfig.json
@@ -1,4 +1,0 @@
-{
-  "extends": "@uploadthing/tsconfig/base",
-  "include": ["src/**/*", "*.ts", "test"]
-}
