--- packages/dropzone/turbo.json
@@ -1,9 +1,0 @@
-{
-  "$schema": "https://turbo.build/schema.json",
-  "extends": ["//"],
-  "tasks": {
-    "build": {
-      "outputs": ["core/**", "react/**", "solid/**", "svelte/**", "vue/**"]
-    }
-  }
-}
