+++ packages/react/package.json
@@ -62,8 +62,8 @@
     "typecheck": "tsc --noEmit"
   },
   "dependencies": {
-    "@uploadthing/dropzone": "workspace:*",
-    "@uploadthing/shared": "workspace:*"
+    "@uploadthing/shared": "workspace:*",
+    "file-selector": "0.6.0"
   },
   "peerDependencies": {
     "next": "*",
