+++ packages/shared/src/index.ts
@@ -5,4 +5,5 @@ export * from "./error";
 export * from "./effect";
 export * from "./component-utils";
 export * from "./crypto";
+export * from "./dropzone-utils";
 export * from "./tagged-errors";
