+++ packages/svelte/package.json
@@ -33,8 +33,8 @@
     "uploadthing": "^7.0.0"
   },
   "dependencies": {
-    "@uploadthing/dropzone": "workspace:*",
-    "@uploadthing/shared": "workspace:*"
+    "@uploadthing/shared": "workspace:*",
+    "file-selector": "0.6.0"
   },
   "devDependencies": {
     "@sveltejs/adapter-auto": "^3.1.1",
