+++ packages/svelte/src/lib/index.ts
@@ -3,5 +3,5 @@ export { UploadButton, UploadDropzone, Uploader } from "./component";
 
 export type * from "./types";
 
-export { createDropzone } from "@uploadthing/dropzone/svelte";
-export type * from "@uploadthing/dropzone/svelte";
+export { createDropzone } from "./component/create-dropzone";
+export type * from "./component/create-dropzone";
