+++ packages/vue/src/index.ts
@@ -3,6 +3,3 @@ export * from "./components/dropzone";
 export * from "./useUploadThing";
 
 export type * from "./types";
-
-export { useDropzone } from "@uploadthing/dropzone/vue";
-export type * from "@uploadthing/dropzone/vue";
